First attempt and actually using the library
Everything seems to be working.  I'd like to make the command line options for
genjwt more user friendly.  But this actually worked surprisingly well and was pretty
straightforward to implement.

The next thing I'd like to do is have the JWT include a URL for some external server
to query.